Lesson 6. Ex4
