+++ basketapp/templates/basketapp/basket.html
@@ -11,7 +11,7 @@
         <span class="product_name">{{ item.product.name }}</span>
         <span class="product_price">{{ item.product.price }}&nbsp ₽</span>
         <input type="number" name="{{ item.pk }}" min="0" value="{{ item.quantity }}">
-        <span class="product_cost">{{ item.cost }}&nbsp ₽</span>
+        <span class="product_cost">{{ item.product_cost }}&nbsp ₽</span>
         <button class="btn btn-round">
             <a href="{% url 'basket:remove' item.pk %}" class="">
                 удалить
@@ -22,8 +22,8 @@
 
     {% if basket_items %}
     <div class="basket_summary">
-        В корзине {{ user.user_basket.total_quantity }} товаров общей стоимостью
-        {{ user.user_basket.total_cost }} руб
+        В корзине {{ user.total_quantity }} товаров общей стоимостью
+        {{ user.total_cost }} руб
     </div>
     {% endif %}
